feat: change EtcdClientFactory to singleton
