+++ eventmesh-registry-plugin/eventmesh-registry-etcd/src/test/java/org/apache/eventmesh/registry/etcd/service/EtcdRegistryServiceTest.java
@@ -73,7 +73,6 @@ public class EtcdRegistryServiceTest {
     public void testInit() {
         etcdRegistryService.init();
         etcdRegistryService.start();
-        Assert.assertNotNull(etcdRegistryService.getServerAddr());
     }
 
     @Test
